feat: added AddBusiness to the db via filling form
