+++ src/components/Business/boostrap-design/assets/css/styles.scss
@@ -1,0 +1,7 @@
+@import '../../../../styles/_variables.scss';
+
+// Restul codului CSS rămâne neschimbat
+.form-container {
+  padding: 20px;
+  // ...
+}
