+++ vite.config.js
@@ -1,8 +1,16 @@
-
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  server: {
+    proxy: {
+      '/api/v1': {
+        target: 'https://estiaproject-b3ef95234cdd.herokuapp.com',
+        changeOrigin: true,
+        secure: false
+      }
+    }
+  }
 })
